requirements.txt
streamlit
streamlit-option-menu
streamlit-echarts
ktrain
altair==4.2.2
spacy>=3.0.0,<4.0.0
en_core_web_lg @ https://github.com/explosion/spacy-models/releases/download/en_core_web_lg-3.4.1/en_core_web_lg-3.4.1-py3-none-any.whl
skillNer
pandas
numpy
ipython
neattext
tensorflow
st_annotated_text==2.0.0
requests
PyPDF2
python-docx

# nltk
# Unidecode
streamlit_lottie